Add a cython test to give perspective on relative ease/performance
The cython implementation is about 2x slower than the Rust
implementation, but it's not clear if the slice in the cython
call is actually doing a copy of the entire memory-area or just
doing a simple reference of the memory.

Note that the cython implementation doesn't build a proper module
with setup.py, it just uses the command-line cythonize operation
to produce a quick-and-dirty extension.